Refactor CleanAirPage.scss, translation.json, and index.js files
