--- website/frontend/src/pages/CleanAir/ForumEvents/index.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Index = () => {
-  return <div>Index</div>;
-};
-
-export default Index;
