+++ website/frontend/src/pages/CleanAir/ForumEventsPages/Partners.js
@@ -1,0 +1,100 @@
+import React from 'react';
+import { Pagination, usePagination } from 'components/CleanAir/pagination/Pagination';
+import { useTranslation, Trans } from 'react-i18next';
+import { SplitTextSection } from 'components/CleanAir';
+
+const ITEMS_PER_PAGE = 6;
+
+const Partners = ({ FundingPartners, OtherPartners }) => {
+  const { t } = useTranslation();
+  // Pagination setup for OtherPartners
+  const { currentItems, currentPage, setCurrentPage, totalPages } = usePagination(
+    OtherPartners || [],
+    ITEMS_PER_PAGE
+  );
+
+  return (
+    <>
+      {FundingPartners && FundingPartners.length > 0 && (
+        <>
+          <section className="Funding_partners">
+            <SplitTextSection
+              lists={[]}
+              content={
+                <div className="partners-wrapper">
+                  <div className="partner-logos">
+                    <div className="grid-container">
+                      {FundingPartners.map((item) => (
+                        <a className="cell" key={item.id} href={item.website_link} target="_blank">
+                          <img
+                            className="logo"
+                            src={item.partner_logo}
+                            alt={item.name}
+                            loading="lazy"
+                          />
+                        </a>
+                      ))}
+                    </div>
+                  </div>
+                </div>
+              }
+              title={
+                <h2 className="section_title">{t('cleanAirSite.Forum.sections.partners.fund')}</h2>
+              }
+              bgColor="#FFFFFF"
+            />
+          </section>
+        </>
+      )}
+
+      {OtherPartners && OtherPartners.length > 0 && (
+        <>
+          <div className="separator" />
+          <section className="forum_partners">
+            <SplitTextSection
+              lists={[]}
+              content={
+                <div style={{}}>
+                  <div className="partners-wrapper">
+                    <div className="partner-logos">
+                      <div className="grid-container">
+                        {currentItems.map((item) => (
+                          <a
+                            className="cell"
+                            key={item.id}
+                            href={item.website_link}
+                            target="_blank">
+                            <img
+                              className="logo"
+                              src={item.partner_logo}
+                              alt={item.name}
+                              loading="lazy"
+                            />
+                          </a>
+                        ))}
+                      </div>
+
+                      <Pagination
+                        currentPage={currentPage}
+                        setCurrentPage={setCurrentPage}
+                        totalPages={totalPages}
+                      />
+                    </div>
+                  </div>
+                </div>
+              }
+              title={
+                <h2 className="section_title">
+                  {t('cleanAirSite.Forum.sections.partners.others')}
+                </h2>
+              }
+              bgColor="#FFFFFF"
+            />
+          </section>
+        </>
+      )}
+    </>
+  );
+};
+
+export default Partners;
